Rudementary Box dragging
Got the box to properly drag in the most basic sense, however it's still really clunky. How it works so far is that whenever we move the mouse just track where it's going (relative to where we clicked on the box) and set the 'top' and 'left' values accordingly. However if you quickly move your mouse outside the box it won't catch it and will stop moving. This is fixed in the next commit.