Better Window Drag
This is better for a few reasons: 1. We make a decision based off a universal state (mouse position) instead of just the Window component. 2. We're utilizing React's 'useLayoutEffect' which handles graphics better than useEffect and stateRerenders. Basically whenever we try to render, we just check if we're in the moving state and if we are, we move the box to the cursor. That simple.