Window Resizing
Got it all to work, this is the flow of actions: 1. Is the mouse near an edge? 2. Which edge is it nearest to? 3. Mouse down? 4. Resize based on where the mouse is. I know it's getting a bit more crowded, but we're getting it running first, and then we'll clean it up.