Added icons and fixed a weird bug
Added the icons and header information (title, minimize, exit) and simply just had them push those dealings to the parent component. There also was a weird bug that when you removed one of the windows from the state, it'd take the attributes and apply it to another window instead of removing that html. My guess is that it's react trying to have as little DOM updates as possible which is fair, but not working so well with my app. I found a work around by just forcing the update for the next targeted element, but I suppose I could go back and fix it more