Start with window item dragging in the Bar
Oh boy, this guy's been a doozy. I thought to maybe just replicated what I did for the window dragging and I did... but there were was a lot more to it. A LOT (as you can maybe tell). Same technique: 1. Set the state to 'dragging' (double the value as an index) 2. Every time the mouse moves and we're in our state, move the window item with it. 3. When we let go of it, then we're done moving. Anyway, I had to do that on items in an array so it got a bit harry. Thankfully it works... decent now. Next is to get it to actually rearrange based on where I'm dragging it