Sliding is working now
Got the sliding to work, whichever left most element you're still to the right of is where you're to be swapped with that index. Sure there are performance problems with it and I'm sure when we introduce components into it there will be problems, but for now she runs. Next we need to get it to not minimize when we're trying to re arrange each item