Cleaned up the code and fixed a bug with minimizing tabs
Ok, I didn't really fix the bug. Really, I traded it in for another, but it's a better bug than the previous one: Dragging to a different spot doesn't always put your application there, but it's fixable with another action which is annoying, but still usable. The other bug, you couldn't even minimize one of the windows. Anyway, besides that, I added comments and formating to the code which is always good ya' know?