Implemented Components with the windows and cleaned up a bit more
So it proved difficult to extract out some logic from the Bar component since everything was so tightly coupled, but I did get the create a clicker and text input component to be used for each window to show that you can input anything you'd want in there and it'd work. There seems to be an overflow issue, but for now that's resolved by changing the min size for items.